TG-110 Specified unused player for win algo tests
In the WinGameResolutionStrategy unit tests, I was passing a player
as if it were necessary, which is confusing for the reader. I decided
to make clear the fact that the player passed as an argument to this
algorithm is unused by it. I made the change also for the
TieGameResolutionStrategy unit tests.