TG-111 Early tie detection algorithm disabled
After recent troubles with the unit testing of this algorithm
(the tests seem fragile), I decided to block its usage until
more tests are done and the confidence is there. In the meantime,
a basic tie detection algorithm is taking its place to make
sure users are not stuck with a completed board with no winner.