TG-114 CommandDropChip logic moved to Model.cpp
The command's logic, I now realize, is too complex to act as a
command. The problem is that the chip drop may not always
succeed (for example if the drop occurs in a colum that
is full) and hence it is possible that no drop happens. In this
case, a call to Undo() will have no effect, since no state
will have changed in the game. This could lead in the user
making multiple calls to undo with no effect.

By bringing back the logic into the model, I will try to rewrite
the command by adding only the successful drop logic in it so
only successful drops will be added to the command stack and
every call to Undo() will undo an actual drop.