TG-168 Redo added for a disc drop
For now, redoing only works once because there is no way to know
(unless an undo has just been done) that redoing is possible. For
that, further services will need to be added to the command stack
and the model. Note that redo has been added to the GUI, but the
'Ctrl + y' binding is not working.