TG-168 Redo bug fixed
The chip was passed as a reference to the drop chip command, and
at some point references got confused. I switched the references
to an owned unique_ptr instead. This means that the command now
owns a copy of the disc in itself, and hence no more corruption
can occur.