TG-185 Reusable logic added for win and tie tests
A new functionnality has been added to the win/tie unit test fixture
to make it possible to quickly add unit tests for the win and tie
algorithms without duplication and the need to code huge tests. The
new logic simply needs an ASCII encoded game, as a string. It then
automatically decodes it and checks the decoded game against the
algorithms.