TG-194 Bug fixed for vertical tie detection
Some cases of the vertical tie detection were detected too early
(when there was not tie). The problem had to do with the algorithm
for computing how much moves left a player had to play did not
take into account moves already completed (and part of a possible
win). This meant that possible wins were overloaked and ties were
detected instead.