TG-194 Extra calls to strategies added to the fixture
Whenever something went wrong in a test and debbuging was needed,
one always needed to go and change the fixture's code to add
extra strategy calls (to be able to break to it) and remove
them after the fix. This was annoying, so the extra calls have
all been added. They add almost no extra overhead (because the
call are not done) when all tests pass, which should pretty
much always be the case after developpement.