TG-174 User help now accessible through Connect X
The "Help -> Contents" menu item now triggers the Gnome help
browser. The top Rule.mk file has been amended to add a
user documentation rule which "installs" the help file on
the system for the Gnome help browser to find. For now, this
"installation" is a hack, which simply copies the help
files to the ~/.local/share destination (or doesn't if
it is not available).

This will make help developpement easier while a build system
is introduced and TG-182 is completed.