TG-237 Strong type mechanics added.
Because animations work using coordinates on a XY plane, a lot of
interfaces make heavy use of doubles to represent various concepts,
such as position, displacements, ratios and so on. A problem with
this is that the interface user must remember which parameter
represents which concept. Furthermore, doubles carry no meaning,
which means that a lot of extra documentation is needed to make
sure the interface is clean.

By using strong types, I beleive that this can be mitigated. The
new animation model is a really good canditate to try it out.