TG-235 Undo and redo work.
Note that when redoing, the same animation as when a chip is dropped
occurs, but since the chip can be anywhere in the NDA, the falling
chip may appear in the wrong column. Once the animation is completed,
the board is refreshed and the chip is "teleported" to the appropriate
location.

This is a minor issue and fixing it would be costly because the
issue comes from the fact that the model sends the same notification
when a chip is dropped and when a redo is requested. They should
be split for this issue to be resolved.