TG-179 Initial CMakeLists.txt files added.
These files were added to test directories. For now, a build can't
happen because header inclusion is not fixed to fit the new format.
Also, a Testing.cmake module was added to centralize some code. It
specifies a function to automatically add a unit test executable
target. This is what all the CMakeLists.txt files for tests are
actually using.