TG-247 Connect X is now partially installable.
The executable and the icons are installable. The user documentation
is not. Also, the log target (which is along the executable file)
is no longer acceptable since most users will install it under
a directory in which they do not have the right to write (under
uer/local, for instance). It need to be reviewed.