TG-240 Bot players added.
To acheive this, the cxmodel::Player type was abstracted. The
cxmodel::IPlayer interface has been added, as well as a factory
function. This made a lot of code need to be updated from the
model to the ui.

The new interface has a `IsManaged` getter function to tell if
it is managed by Connect X (bot), or not (human).