TG-244 Animation strategy added.
For now, the animation is still hooked to the right arrow key, so
the game is not playable. The strategy has been added and some
logic was added to the animated board to handle both types of
right moves, especially in the Redraw, which assumed only one move
type per direction.