TG-262 First player now always non-managed.
If the first player is managed, a display bug occurs which make it
impossible to see the animation. The game stars with n tokens already
in place. Worst is that if the firts n players are managed, the first
n tokens are placed one on top of the other, making it look ridiculous.
For now, I a patching this by making the first player non managed,
but later on this should be made possible.