TG-263 Command stack now executes a new command first.
Before, `CommandStack::Execute` included the command in its
command list and then executed it (from the list). Now that
we want to add a status returned from `ICommand::Execute`,
we can no longer do that because if the command fails, it will
still be in the list. So a new command now executes and then
is included in the list (no error handling for now).

Note that the `CommandDropChip` started to misbehave because
it was handling a notification inside the command, which led
to the stack not being updated as the GUI was (the command was
added to the stack after the GUI had been updated). This was
also fixed by moving the "drop chip" notification out of the
command. The "redo drop" is still there and works well. It will
be handled later, if it becomes a necessity.