Assertion in unit tests silenced.
The command that creates new games cannot be undone. For nos, trying
to do so throws an assertion. Such an assertion was printed in the
unit test report for the test making sure nothing happened (in the
eyes of the user) if undo was called. We are now testing that
the assertion occurs by cathing it, and it no longer appears in
the logs.