TG-303 Undo on first drop now possible.
About the impossibility to undo, the issue was that the first chip
drop was executed from the command stack and before it has a chance
to complete (from the point of view of the command stack), it notifies
its observer views that it has completed. The Main window reacts and
checks if undo can be done. Since the drop command is not yet
completed, it is not yet added to the command stack and the Main
window thinks it cannot unblock undoes. I added an additional check
to avoid this.

I also found an issue in the Game view with the disabling/enabling
of the key handlers. There was a mismatch when redoing/undoing
because the keyboard shortcuts for those use a key modifier (Ctrl).
I noticed that when pressing a key modifier alone, a key pressed
event is sent, as if it were a normal key. This is what caused the
mismatch.