TG-310 The presenter now drives the color combo boxes.
It is now the New Game view presenter that supplies the list of
colors to show in the color combo boxes that let the user pick a
chip color.

With this change, It will be possible to add logic to the presenter
to show only the relevent colors and avoid the user possible pick
errors.