TG-315 Game drawing surface cached.
The whole board surface (i.e. everything that is drawn in the
animated board widget) is cached in memory and only created once,
at the first drawn frame. On each subsequent frame drawing, the
surface is cleared, instead of destroyed and the new frame is
drawn over it. This avoids constantly re-creating a new surface
for every frame, which seems to be expensive according to callgrind.

The new callgring run seems to indicate the cost is now lower. Note
that this only is a test integration: the current implementation does
not handle resizing the window. This will come later, is this commit
proves to be relevent.