TG-350 Menu items partially abstracted.
The previous way of doing things did not managage memory appropriately
since pointers were deleted at the end of the factory function. To
avoid these lifetime management issues, a type was created to represent
menu items. It is not yet abstracted away completely (it will be done
in TG-355), but made possible the usage of cxgui::ILayout in the
menu item implementation.