TG-360 Delegate now used.
While testing, I noticed that removing a row was causing a crash.
This is becuase the switch is a `std::unique_ptr` and that the
`cxgui::NewPlayerRow` stores a raw pointer in a `Gtk::Grid`. When
`cxgui::Gtkmm3NewPlayersList::RemovedManaged` is called, `delete`
is called on the raw pointer stored in the `Gtk::Grid` bypassing
the `std::unique_ptr`. When the `std::unique_ptr`'s deleter is called,
a crash occures because `delete` was already called.

The solution (for now) was to have the switch as a `std::shared_ptr`,
which makes sure more than one `delete` can be performed.