TG-360 `cxgui::Gtkmm3NewPlayersList` simplified.
In the `cxgui::Gtkmm3NewPlayersList` class, `Gtk::managed` was
used to handle row memory, which was causing conflict (and UB)
while the different containers were deleted (e.g. when a row
was deleted). The data is now completely owned by the
`cxgui::Gtkmm3NewPlayersList` class, which simplifies my porting
to my own containers.