TG-380 cxstd/ --> cxcmn/
The `cxstd` library is renamed `cxcmn`. This is because using the
`cx::std` namespace caused name collisions with the standard namespace.
At this point, the library has been renamed as far as CMake is
concerned. It has also been renamed on disk. However, the `cxstd`
namespace is still used in the code. It will be migrated in the
next commin.