TG-380 Added documentation for the `cxcmnui` library.
This high level documentation explain the main ideas behind the
design which takes advantage of the Abstract Factory pattern. A
"Hello World" example has been added, but does not yet build
because of issues found while documenting. These will be fixed
in the next commits.

To add this documentation, a new "doxygen" repository has been
added. it includes all of the extra documentation which is not
directly in the code.